:sparkles: Add DropdownItems to Near Amount Account Component
